bug tests

+++ app/src/main/java/com/protheansoftware/gab/model/Profile.java
@@ -6,7 +6,7 @@ package com.protheansoftware.gab.model;
 public class Profile {
     private final String name;
     private final int id;
-    private final int fbId;
+    private final long fbId;
 
     public Profile(String name, int id, int fbId){
         this.name = name;
@@ -22,7 +22,7 @@ public class Profile {
         return name;
     }
 
-    public int getFbId() {
+    public long  getFbId() {
         return fbId;
     }
 }
